refactor: move notification related servies to Notification namespace
