+++ src/App/VRChatContentPublisher.App/Extensions/ServicesExtenstion.cs
@@ -2,7 +2,8 @@
 using VRChatContentPublisher.App.Services;
 using VRChatContentPublisher.App.Services.AppLifetime;
 using VRChatContentPublisher.App.Services.Dialog;
-using VRChatContentPublisher.App.Services.NotificationSender;
+using VRChatContentPublisher.App.Services.Notification;
+using VRChatContentPublisher.App.Services.Notification.Sender;
 using VRChatContentPublisher.App.Services.Update;
 using VRChatContentPublisher.App.ViewModels;
 using VRChatContentPublisher.App.ViewModels.Data;
@@ -31,7 +32,7 @@ public static class ServicesExtenstion
 
         services.AddSingleton<AppLifetimeService>();
         services.AddSingleton<AppWindowService>();
-        services.AddSingleton<AppNotificationService>();
+        services.AddSingleton<DesktopNotificationService>();
         services.AddSingleton<IActivateWindowService>(s => s.GetRequiredService<AppWindowService>());
 
         services.AddSingleton<InAppNotificationService>();
